perf(course): 优化课程删除逻辑
- 为 deleteCourseByIds 方法添加 @Transactional 注解，确保数据一致性
- 在删除课程前增加判断，避免出现 SQL 语句执行出错的情况
- 添加对 CoursePerformer 表中相关数据的删除，确保数据完整性

+++ src/main/java/com/sqx/modules/course/service/impl/CourseServiceImpl.java
@@ -462,12 +462,15 @@ public class CourseServiceImpl extends ServiceImpl<CourseDao, Course> implements
     }
 
 
+    @Transactional
     @Override
     public Result deleteCourseByIds(String ids) {
         for (String id : ids.split(",")) {
-            baseMapper.deleteById(Long.parseLong(id));
-            courseDetailsDao.delete(new QueryWrapper<CourseDetails>().eq("course_id", Long.parseLong(id)));
-            redisUtils.deleteByPattern(String.format("*%s*", id));
+            if (baseMapper.deleteById(Long.parseLong(id)) > 0) {
+                coursePerformerService.remove(new QueryWrapper<CoursePerformer>().eq("course_id", Long.parseLong(id)));
+                courseDetailsDao.delete(new QueryWrapper<CourseDetails>().eq("course_id", Long.parseLong(id)));
+                redisUtils.deleteByPattern(String.format("*%s*", id));
+            }
         }
         redisUtils.deleteByPattern("page*");
         return Result.success();
